fixed things that annoyed me
Disbaled the automatic reorganisation of slps into seperate folders. 
This feature currently causes a crash when the tournament name is set in the dashboard and also seems to be linked to randomly failing stat generation.

The crash is caused by an unhandled IO exception, that happens when one of the libs tries to access the slp file on it's original path. 
Unsure about the stat generation issue, but when testing stat generation didn't faile if the slp file wasn't moved.